diversion: Refactoring!!!
